requirements.txt
ftfy==6.2.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pytesseract==0.3.10
wcwidth==0.2.13
